updated project structure and logo added
