--- Term_6-Web/io.game/index.html
@@ -1,8 +1,0 @@
-<style>
-    body {
-        margin: 0;
-        background-color: 0;
-    }
-</style>
-<canvas></canvas>
-<script src="index.js"></script>
