+++ Term_6-Web/io.game/src/client/html/index.html
@@ -7,4 +7,4 @@
 <canvas></canvas>
 
 <script src="/socket.io/socket.io.js"></script>
-<script type="module" src="index.js"></script>
+<script src="./game.83fe12bdd000fb02aae4.js"></script>
