+++ Term_6-Web/io.game/src/server/object.js
@@ -1,5 +1,5 @@
 const Vector = require('./vector')
-const Constants = require('./constants')
+const Constants = require('../shared/constants')
 
 class Object {
     constructor(id, x, y){
