+++ Term_6-Web/io.game/src/server/player.js
@@ -1,5 +1,5 @@
 const MovableObject = require('./movableobject')
-const Constants = require('./constants')
+const Constants = require('../shared/constants')
 
 class Player extends MovableObject {
     constructor(id, username, x, y, r) {
