+++ Term_6-Web/io.game/src/server/room.js
@@ -1,4 +1,4 @@
-const Constants = require('./constants')
+const Constants = require('../shared/constants')
 const Player = require('./player')
 const Bullet = require('./bullet')
